backend/requirements.txt
# =========================
# Core Backend
# =========================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2

# =========================
# LangChain Ecosystem (FIXED)
# =========================
langchain==0.3.7
langchain-community==0.3.7

# LLM Providers
langchain-google-genai==2.0.4
langchain-openai==0.2.6
langchain-anthropic==0.3.0
langchain-ollama==0.2.1

# =========================
# Google AI / Gemini
# =========================
google-generativeai==0.8.3

# =========================
# RAG (Vector DB + Embeddings)
# =========================
faiss-cpu==1.9.0
sentence-transformers==3.3.1
langchain-huggingface==0.1.2

# =========================
# Document Processing
# =========================
PyPDF2==3.0.1
python-docx==1.1.2
pypdf==5.1.0

# =========================
# Tokenization
# =========================
tiktoken==0.8.0

# =========================
# Search Tool
# =========================
tavily-python==0.5.0

# =========================
# Vision / Image Processing
# =========================
opencv-python-headless==4.10.0.84
Pillow==11.0.0